requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-jwt-extended==4.3.1
python-dotenv==0.19.0
Werkzeug==2.0.3
psycopg2-binary==2.9.3
yfinance==0.2.31
requests==2.31.0
pandas>=1.3.0